# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Sports Auction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments/{tournamentId}/analytics/summary:
    get:
      operationId: getTournamentSummary
      tags: [analytics]
      summary: Get high-level auction summary stats
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentSummary"
  /tournaments/{tournamentId}/analytics/team-purses:
    get:
      operationId: getTeamPurses
      tags: [analytics]
      summary: Get purse remaining and spent per team
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team purse breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamPurse"
...
  /tournaments/{tournamentId}/analytics/top-bids:
    get:
      operationId: getTopBids
      tags: [analytics]
      summary: Get top sold players by price
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Top sold players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopBidEntry"
  /tournaments/{tournamentId}/analytics/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [analytics]
      summary: Get sold/unsold counts per category
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

  # ─── GLOBAL PLAYER DATABASE ───────────────────────────────────────────────────
...
components:
  schemas:
...
    TournamentSummary:
      type: object
      required: [totalPlayers, soldPlayers, retainedPlayers, unsoldPlayers, availablePlayers, totalSpent, totalTeams, avgBidAmount]
      properties:
        totalPlayers:
          type: integer
        soldPlayers:
          type: integer
        retainedPlayers:
          type: integer
        unsoldPlayers:
          type: integer
        availablePlayers:
          type: integer
        totalSpent:
          type: integer
        totalTeams:
          type: integer
        avgBidAmount:
          type: integer
        highestBid:
          type: integer
        mostActiveTeam:
          type: ["string", "null"]
    TeamPurse:
      type: object
      required: [teamId, teamName, shortCode, color, originalPurse, boosterTotal, effectiveCapacity, purse, purseUsed, purseRemaining, playersBought, ownerName, reservePurse, spendablePurse, slotsRequired, lowestBasePrice, minimumSquadSize, maximumSquadSize, retainedCount]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        shortCode:
          type: string
        ownerName:
          type: string
        color:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        originalPurse:
          type: integer
          description: Immutable baseline purse from team setup
        boosterTotal:
          type: integer
          description: Sum of active purse boosters
        effectiveCapacity:
          type: integer
          description: originalPurse + boosterTotal
        purse:
          type: integer
          description: Alias for effectiveCapacity (backward compatible)
        purseUsed:
          type: integer
        purseRemaining:
          type: integer
        playersBought:
          type: integer
        reservePurse:
          type: integer
        spendablePurse:
          type: integer
        slotsRequired:
          type: integer
        lowestBasePrice:
          type: integer
        minimumSquadSize:
          type: integer
        maximumSquadSize:
          type: integer
        retainedCount:
          type: integer
        topPlayerName:
          type: ["string", "null"]
        topPlayerAmount:
          type: ["integer", "null"]
    TopBidEntry:
      type: object
      required: [playerId, playerName, teamName, soldPrice, basePrice, categoryName]
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        teamName:
          type: string
        teamColor:
          type: ["string", "null"]
        soldPrice:
          type: integer
        basePrice:
          type: integer
        categoryName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
...
    CategoryBreakdown:
      type: object
      required: [categoryId, categoryName, total, sold, unsold, available]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        colorCode:
          type: ["string", "null"]
        total:
          type: integer
        sold:
          type: integer
        unsold:
          type: integer
        available:
          type: integer
        totalSpent:
          type: integer

    # ─── Global Player Database schemas ───────────────────────────────────────
...